Refactors dna
Currently, dna is created as a string, which is both inefficient
and impractical. This commit changes it to be an array of objects.

It still has to be converted to a string to check if it's unique,
and I've removed `bypassDNA` to maintain compatibility with the
main repo because the dna produced from there isn't filtered,
meaning otherwise I wouldn't be able to check whether dna created
in this is equal to that created in the main repo and vice versa.